Refactor LanguageResources
